permission control on post create, update, and delete

+++ posts/views.py
@@ -14,6 +14,8 @@ def post_home(request):
 # def post_create(request):
 #     return HttpResponse("<h1> Create </h1>")
 def post_create(request):
+    if not (request.user.is_staff or request.user.is_superuser):
+        raise Http404
     form = PostForm(request.POST or None , request.FILES or None)
     if form.is_valid():
         form.save()
@@ -35,8 +37,10 @@ def post_detail(request, post_slug):
     "share_string": quote(instance.content)
     }
     return render(request, 'post_detail.html', context)
-    
+
 def post_update(request, post_slug):
+    if not (request.user.is_staff or request.user.is_superuser):
+        raise Http404
     instance = get_object_or_404(Post, slug=post_slug)
     form = PostForm(request.POST or None, request.FILES or None, instance = instance)
     if form.is_valid():
@@ -51,6 +55,8 @@ def post_update(request, post_slug):
     return render(request, 'post_update.html', context)
 
 def post_delete(request, post_slug):
+    if not (request.user.is_staff or request.user.is_superuser):
+        raise Http404
     instance = get_object_or_404(Post, slug=post_slug)
     instance.delete()
     messages.success(request, "Successfully Deleted!")
